requirements.txt
python-telegram-bot==21.6
gspread==6.1.0
oauth2client==4.1.3
openai>=1.40.0
requests
flask